slimmer wasm bundle
