+++ src/lib.rs
@@ -28,9 +28,15 @@
 //! A library for turning computational graphs, such as neural networks, into ZK-circuits.
 //!
 
+use circuit::Tolerance;
+use clap::Args;
+use graph::Visibility;
+use serde::{Deserialize, Serialize};
+
 /// Methods for configuring tensor operations and assigning values to them in a Halo2 circuit.
 pub mod circuit;
 /// CLI commands.
+#[cfg(not(target_arch = "wasm32"))]
 pub mod commands;
 #[cfg(not(target_arch = "wasm32"))]
 #[allow(missing_docs, reason = "abigen doesn't generate docs for this module")]
@@ -38,6 +44,7 @@ pub mod commands;
 pub mod eth;
 /// Command execution
 ///
+#[cfg(not(target_arch = "wasm32"))]
 pub mod execute;
 /// Utilities for converting from Halo2 Field types to integers (and vice-versa).
 pub mod fieldutils;
@@ -46,6 +53,7 @@ pub mod fieldutils;
 #[cfg(feature = "onnx")]
 pub mod graph;
 /// beautiful logging
+#[cfg(not(target_arch = "wasm32"))]
 pub mod logger;
 /// Tools for proofs and verification used by cli
 pub mod pfsys;
@@ -57,3 +65,32 @@ pub mod tensor;
 /// wasm prover and verifier
 #[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
 pub mod wasm;
+
+/// Parameters specific to a proving run
+#[derive(Debug, Copy, Args, Deserialize, Serialize, Clone, Default, PartialEq, PartialOrd)]
+pub struct RunArgs {
+    /// The tolerance for error on model outputs
+    #[arg(short = 'T', long, default_value = "0")]
+    pub tolerance: Tolerance,
+    /// The denominator in the fixed point representation used when quantizing
+    #[arg(short = 'S', long, default_value = "7")]
+    pub scale: u32,
+    /// The number of bits used in lookup tables
+    #[arg(short = 'B', long, default_value = "16")]
+    pub bits: usize,
+    /// The log_2 number of rows
+    #[arg(short = 'K', long, default_value = "17")]
+    pub logrows: u32,
+    /// The number of batches to split the input data into
+    #[arg(long, default_value = "1")]
+    pub batch_size: usize,
+    /// Flags whether inputs are public, private, hashed
+    #[arg(long, default_value = "private")]
+    pub input_visibility: Visibility,
+    /// Flags whether outputs are public, private, hashed
+    #[arg(long, default_value = "public")]
+    pub output_visibility: Visibility,
+    /// Flags whether params are public, private, hashed
+    #[arg(long, default_value = "private")]
+    pub param_visibility: Visibility,
+}
